* refactor-test:
  properly align the slightly larger DAG
  special test for the bug we were encoutering
  remove print statements probably left over from debugging
  potentially fix all the issues by removing unneeded code
  fix aligment of all examples
  display what we get
  fix alignment in test example
  the tricky bug and what has to be done about it
  add a debug mode for testing
  initial test for something more complex
  test utility for printing a dict so that its readable
  test utility for tagging
  add test for tree-tags
  simplify setup
  use shlex to parse commands
  basic test for blob-tags
  test a linear history with a single tag
  slightly change test-case semantics and graph-layout

+++ test/test.py
@@ -26,9 +26,15 @@ import shutil as sh
 import unittest as ut
 import git_big_picture as gbp
 import git_big_picture.git_tools as gt
+import shlex
+
+debug=False
 
 def dispatch(command_string):
-	return gt.get_command_output(command_string.split(' '))
+	return gt.get_command_output(shlex.split(command_string))
+
+def tag(sha1, tag_name):
+	dispatch('git tag %s %s' % (tag_name, sha1))
 
 def get_head_sha():
 	return dispatch('git rev-parse HEAD').rstrip()
@@ -36,6 +42,10 @@ def get_head_sha():
 def empty_commit(mess):
 	dispatch('git commit --allow-empty -m %s' % mess)
 	return get_head_sha()
+
+def print_dict(dict_):
+	for (k,v) in dict_.iteritems():
+		print k,v
 
 class TestGitTools(ut.TestCase):
 
@@ -46,44 +56,43 @@ class TestGitTools(ut.TestCase):
 
 		"""
 		self.testing_dir = tf.mkdtemp(prefix='gbp-testing-', dir="/tmp")
+		if debug:
+			print self.testing_dir
 		gbp.git_tools.git_env = {'GIT_DIR' : "%s/.git" % self.testing_dir }
-		gt.get_command_output(['git', 'init', self.testing_dir])
 		self.oldpwd = os.getcwd()
 		os.chdir(self.testing_dir)
 
-		dispatch('/usr/bin/git init')
+		dispatch('git init')
 		dispatch('git config user.name git-big-picture')
 		dispatch('git config user.email git-big-picture@example.org')
 
 	def tearDown(self):
 		""" Remove testing environment """
-		sh.rmtree(self.testing_dir)
+		if not debug:
+			sh.rmtree(self.testing_dir)
 		os.chdir(self.oldpwd)
 
 	def test_get_parent_map(self):
 		""" Check get_parent_map() works:
 
-				 4 <- other
-				 |\
-				 | \
-	   master -> 2  3
-				 | /
-				 |/
-				 1
-
+		    master other
+		        |   |
+		    A---B---D
+		     \     /
+		      --C--
 		"""
-		sha_1 = empty_commit('1')
-		sha_2 = empty_commit('2')
+		a = empty_commit('a')
+		b = empty_commit('b')
 		dispatch('git checkout -b other HEAD^')
-		sha_3 = empty_commit('3')
+		c = empty_commit('c')
 		dispatch('git merge --no-ff master')
-		sha_4 = get_head_sha()
+		d = get_head_sha()
 
 		expected_parents = {
-			sha_1:set(),
-			sha_2:set((sha_1,)),
-			sha_3:set((sha_1,)),
-			sha_4:set((sha_2, sha_3)),
+			a:set(),
+			b:set((a,)),
+			c:set((a,)),
+			d:set((c, b)),
 		}
 
 		actual_parents = gt.get_parent_map()
@@ -92,24 +101,232 @@ class TestGitTools(ut.TestCase):
 
 	def test_remove_non_labels_one(self):
 		""" Remove a single commit from between two commits.
-			3 <-- master
-			2
-			1 <-- one
+
+		    A---B---C
+		    |       |
+		   one    master
 
 		No ref pointing to B, thus it should be removed.
 
 		"""
-		one = empty_commit('1')
+		a = empty_commit('A')
 		dispatch('git branch one')
-		two = empty_commit('2')
-		three = empty_commit('3')
+		b = empty_commit('B')
+		c = empty_commit('C')
 		(lb, rb, ab), (tags, ctags, nctags) = gt.get_mappings()
 		graph = gbp.CommitGraph(gt.get_parent_map(), ab, tags)
 		graph._remove_non_labels()
 		expected_reduced_parents = {
-			one:set(),
-			three:set((one,)),
+			a:set(),
+			c:set((a,)),
+		}
+		self.assertEqual(expected_reduced_parents, graph.parents)
+	
+	def test_remove_non_labels_with_tags(self):
+		""" Remove three commits and root commmit
+
+            A---B---C---D---E---F
+                |               |
+               0.1            master
+
+		"""
+		a = empty_commit('A')
+		b = empty_commit('B')
+		dispatch('git tag 0.1')
+		c = empty_commit('C')
+		d = empty_commit('D')
+		e = empty_commit('E')
+		f = empty_commit('F')
+		(lb, rb, ab), (tags, ctags, nctags) = gt.get_mappings()
+		graph = gbp.CommitGraph(gt.get_parent_map(), ab, tags)
+		graph._remove_non_labels()
+		expected_reduced_parents = {
+			b:set(),
+			f:set((b,)),
 		}
 		self.assertEqual(expected_reduced_parents, graph.parents)
 
+	def test_no_commit_tags(self):
+		""" Test for tree-tag and a blob-tag.
+		"""
+
+		a = empty_commit('A')
+		f = open('foo','w')
+		f.writelines('bar')
+		f.close()
+		blob_hash = dispatch('git hash-object -w foo').rstrip()
+		dispatch('git tag -m "blob-tag" blob-tag '+blob_hash)
+		os.mkdir('baz')
+		f = open('baz/foo','w')
+		f.writelines('bar')
+		f.close()
+		dispatch('git add baz/foo')
+		tree_hash = dispatch('git write-tree --prefix=baz').rstrip()
+		dispatch('git tag -m "tree-tag" tree-tag '+tree_hash)
+		dispatch('git reset')
+
+		(lb, rb, ab), (tags, ctags, nctags) = gt.get_mappings()
+		graph = gbp.CommitGraph(gt.get_parent_map(), ab, tags)
+		graph._remove_non_labels()
+		expected_reduced_parents = {
+			blob_hash:set(),
+			tree_hash:set(),
+			a:set(),
+		}
+		self.assertEqual(expected_reduced_parents, graph.parents)
+
+	def test_parent_of_parent_loop(self):
+		""" Test the case, where an alternative route may lead to a parents
+		parent.
+
+		   0.1         0.2    master
+		    |           |       |
+		    A---B---C---D---E---F
+		             \     /
+		              --G--
+
+		   0.1 0.2 master
+		    |   |   |
+		    A---D---F
+		    \      /
+		     ------
+
+		"""
+		a = empty_commit('A')
+		tag(a, '0.1')
+		b = empty_commit('B')
+		c = empty_commit('C')
+		d = empty_commit('D')
+		tag(d, '0.2')
+		e = empty_commit('E')
+
+		dispatch('git checkout -b topic %s' % c)
+		g = empty_commit('G')
+		dispatch('git checkout master')
+		dispatch('git merge topic')
+		f = get_head_sha()
+		dispatch('git branch -d topic')
+
+		(lb, rb, ab), (tags, ctags, nctags) = gt.get_mappings()
+		graph = gbp.CommitGraph(gt.get_parent_map(), ab, tags)
+		graph._remove_non_labels()
+		expected_reduced_parents = {
+			d:set((a,)),
+			a:set(),
+			f:set((a, d,)),
+		}
+		self.assertEqual(expected_reduced_parents, graph.parents)
+
+	def test_expose_multi_parent_bug(self):
+		""" Test for a peculiar bug in pruning the graph.
+
+		Observe the following graph:
+
+			 A---B---C---D---E---F master
+			 |   |    \         /
+			0.0 0.1    N---O---P topic
+
+		It should be:
+
+			0.0---0.1---master
+			        \     /
+			         topic
+
+		But it is:
+
+			0.0---0.1-topic-master
+
+		"""
+		a = empty_commit('A')
+		tag(a, '0.0')
+		b = empty_commit('B')
+		tag(b, '0.1')
+		c = empty_commit('C')
+		d = empty_commit('D')
+		e = empty_commit('E')
+		dispatch('git checkout -b topic %s' % c)
+		n = empty_commit('N')
+		o = empty_commit('O')
+		p = empty_commit('P')
+		dispatch('git checkout master')
+		dispatch('git merge topic')
+		f = get_head_sha()
+		(lb, rb, ab), (tags, ctags, nctags) = gt.get_mappings()
+		graph = gbp.CommitGraph(gt.get_parent_map(), ab, tags)
+		graph._remove_non_labels()
+		expected_reduced_parents = {
+			b:set((a,)),
+			a:set(),
+			f:set((p, b,)),
+			p:set((b,)),
+		}
+		print "a", a
+		print "b", b
+		print "p", p
+		print "f", f
+		print dispatch("git log --oneline %s..%s" % (f, p))
+		print_dict(expected_reduced_parents)
+		print_dict(graph.parents)
+		self.assertEqual(expected_reduced_parents, graph.parents)
+
+	def more_realistic(self):
+		""" Test a slightly larger DAG
+
+		input:
+			        0.1.1   0.1.2
+			          |       |
+			0.0   G---H---I---J---K---L---M maint
+			|    /
+			A---B---C---D---E---F master
+			    |    \         /
+			   0.1    N---O---P topic
+
+		output:
+
+			         0.1.1---0.1.2---maint
+			        /
+			0.0---0.1---master
+			        \     /
+			         topic
+		"""
+		a = empty_commit('A')
+		tag(a, '0.0')
+		b = empty_commit('B')
+		tag(b, '0.1')
+		c = empty_commit('C')
+		d = empty_commit('D')
+		e = empty_commit('E')
+		dispatch('git checkout -b maint %s' % b)
+		g = empty_commit('G')
+		h = empty_commit('H')
+		tag(h, '0.1.1')
+		i = empty_commit('I')
+		j = empty_commit('J')
+		tag(j, '0.1.2')
+		k = empty_commit('K')
+		l = empty_commit('L')
+		m = empty_commit('M')
+		dispatch('git checkout -b topic %s' % c)
+		n = empty_commit('N')
+		o = empty_commit('O')
+		p = empty_commit('P')
+		dispatch('git checkout master')
+		dispatch('git merge topic')
+		f = get_head_sha()
+
+		(lb, rb, ab), (tags, ctags, nctags) = gt.get_mappings()
+		graph = gbp.CommitGraph(gt.get_parent_map(), ab, tags)
+		graph._remove_non_labels()
+		expected_reduced_parents = {
+			m:set((j,)),
+			j:set((h,)),
+			h:set((b,)),
+			b:set((a,)),
+			a:set(),
+			f:set((p, b,)),
+			p:set((b,)),
+		}
+		print_dict(expected_reduced_parents)
+		print_dict(graph.parents)
+		self.assertEqual(expected_reduced_parents, graph.parents)
 # vim: set noexpandtab:
